sede service

+++ src/main/java/com/idat/citaslimatambo/service/SedeService.java
@@ -7,9 +7,9 @@ import com.idat.citaslimatambo.model.Sede;
 public interface SedeService {
 	public List<Sede> getAll();
 	 
-	public Sede getById(long id);
+	public Sede getById(int id);
 	 
 	public void saveOrUpdate(Sede sede);
 	 
-	public void delete(long id);
+	public void delete(int id);
 }
